Fix byte-order issues with known field interpretation.
The old code would only work when compiled for little-endian
architectures.  Add new functions convert_uint8(), convert_uint16()
and convert_int16() to convert the values in a CPU endian-agnostic
way.  The return value depends only on the encoded data byte-order,
which is little-endian for IDC images.

All functions are inlined.  convert_uint8() is only for consistency.