Fix order of libraries in Makefile.
The override for LDLIBS specified in the Makefile appended the
required -l flags to any previously given on the command line.  This
does not work in most cases, as additional libraries required for a
specific platform may have to be specified _after_ those explicitly
required by the project.  This change switches the order.