Add configure option to use external libcintelhex.
Introduce --with-cintelhex option to let the user provide an
installation prefix to check.  Also accept "internal" to use the
embedded copy, or "no" to disable.  The default is to auto-detect a
system-wide installation, then fall back to the embedded copy or
ultimately exclude Intel Hex support if all tests failed.