Update to newer libcintelhex release.
The build system for libcintelhex has been dramatically improved and
is now using GNU automake and GNU libtool.  This makes it a lot easier
to build this package with an embedded copy, including support for
VPATH builds.

Adjust the library location for the new libtool path.  The elf-mangle
package does not use libtool yet.