Simplify libcintelhex linker flag handling.
Use AC_SEARCH_LIBS in place of AC_CHECK_LIB macro for the benefit of
automatically adding the needed -l flag to the @LIBS@ substitution.

For the case of using the embedded libcintelhex copy, list the static
library file directly instead of providing an -lcintelhex flag.
Remove the corresponding manual LDFLAGS and LDADD computations.