Rename option byte-size to field-size.
The term "byte size" is ambiguous and could as well describe the whole
image size or the size of a byte.  The new name is more specific.

Replace the old option name completely, since backward compatibility
is currently no big concern and also switch the short option from "-b"
to "-F".