Use AC_SEARCH_LIBS for libelf and improve error messages.
Allow specifying an alternate library name (e.g. -lelf-freebsd on
Debian GNU/kFreeBSD) instead of -lelf.

This is the result of some experimentation with pkg-config's
PKG_CHECK_MODULES macro, which in the end seemed to cause more
problems than it solves.  Instead, we rather point to pkg-config as a
helper to find the needed settings for the standard flags variables.

Even using PKG_CHECK_MODULES conditionally only when defined creates
more headaches.  It uses separate influential environment variables,
making it harder to give helpful hints in the error message.