Add additional lpstrings command line options.
If either --output-separator (-s) or --radix (-t) is given, switch to
the simpler output format from the POSIX specification for strings(1).

As an extension, print the string length in bytes without NUL
terminator for each string if the --length option is given in addition
to --radix.  It is appended after the offset with a plus sign and uses
the same radix.