Add an example implementation of custom post-processor functions.
Just as with the custom options and known fields, this is hidden
behind a configure switch and otherwise falls back to the dummy
implementation.

The example includes a CRC-32 calculation over the whole image, except
for a special field which holds the CRC value itself.  Two
post-processors are defined to either just verify or also update the
CRC value.