Fix erroneous check for missing description.
print_symbol_description_iterator() uses the symbol name itself as a
fallback for the field description.  However, the check was looking at
whether a print function was configured.  This could theoretically
lead to a NULL pointer dereference.  In practice, fields without a
description are in the "unknown" set, so no print function will be
configured either.  Thus the bug had no practical relevance.