Provide a field for storing a symbol's "original" value.
The elf-mangle is built to mangle or alter the contents of a binary
image based on symbol definitions.  For some features, it will be good
to keep the original value read from the ELF map file though, e.g. to
detect changes made by the override options or overwritten from the
input image.

Give the symbol structure an additional pointer field where a copy of
the original data can be referenced.  This memory should be allocated
from the heap and must be freed again after use.  Provide the
convenience function symbol_list_free() to accomplish that and set the
respective member to NULL afterwards.