Save original symbol values if requested.
Add an argument save_values to symbol_map_parse() and if enabled, make
a copy of each symbol's content when generating the symbol list.  For
the input map, this function is controlled through the
showFilterChanged flag, and thus symbol_list_free() called for
cleanup.  For the output map, the feature is always disabled, thus
freeing is never necessary.