Switch to a separate command line option for showFilterChanged.
It's perfectly useful to combine the filter option even with other
forms of output, to get a quick glance about what changed between a
map file and the corresponding image.  Defining format variants for
the --print= option for each of them gets bloated, so instead
introduce a --changed option to simply flip on the filter flag.

Fix the enumeration constant for printDefines.  It's not used as flag
bits, but a classic enumeration.