Fix erroneously not transferred fields from source to target map.
When looking up whether fields are the same in the source and target
layout map, field_list_find() is called on the fields_unknown field
list.  These however are not sorted alphabetically, so the
optimization to break early fails to actually find the matching input
field although it exists.  This is usually not a problem with known
fields, since they are sorted alphabetically in the list definition.

Simply remove the optimization relying on a sorted list, which
increases the worst-case run time for every first encounter of a
symbol.