Fix --changed option filtering with ELF output layout map.
In case a second ELF file was specified as output layout map, checking
for differences should happen relative to that map.  But the original
values were not saved at all, so no differences could actually be
checked.

Pass the same configuration flag check to symbol_map_parse() in both
invocations and properly use symbol_list_free() on it as well if
required.