Fall back to hex-dump when pretty-printing unknown fields.
For pretty-print mode, fields without a designated print function
previously were only listed with their description or symbol name.  If
the corresponding field is actually known, but has a NULL entry for
its print function, that may be intended.  For unknown fields however,
it means they show up in a pretty-print listing with their symbol, but
never display a formatted value.

Just like we fall back to printing the symbol name if there is no
field description, it's more useful to dump the raw byte contents
instead of skipping the value altogether.