Simplify elf_errmsg() calls.
The elf_errmsg() function guarantees to return a non-NULL value if
called with the -1 argument, while it may return NULL if there is no
reported error.  That shouldn't happen here, because elf_errno() is
always checked right after a failure return value.  But relying on the
success string fallback with the designated -1 argument seems more
robust.  This doesn't reset the error code anymore as elf_errno() did,
but that shouldn't hurt because it's always within the failed API call
branch.